Move widget data in their own store
