+++ libs/search-widget/src/core/old-stores/main.store.ts
@@ -1,4 +1,4 @@
-import type { DisplayedResource, Intents, WidgetAction } from '../models';
+import type { DisplayedResource, Intents } from '../models';
 import { NO_RESULTS, PENDING_RESULTS } from '../models';
 import { getLabels } from '../api';
 import {
@@ -14,13 +14,7 @@ import {
   switchMap,
   tap,
 } from 'rxjs';
-import type { Classification, IResource, Labels, Search, SearchOptions, Widget } from '@nuclia/core';
-
-let widgetActions: WidgetAction[] = [];
-export const setWidgetActions = (actions: WidgetAction[]) => {
-  widgetActions = actions;
-};
-export const getWidgetActions = () => widgetActions;
+import type { Classification, IResource, Labels, Search, SearchOptions } from '@nuclia/core';
 
 type NucliaStore = {
   query: BehaviorSubject<string>;
@@ -30,7 +24,6 @@ type NucliaStore = {
   searchResults: BehaviorSubject<Search.Results | typeof PENDING_RESULTS>;
   triggerSearch: Subject<void>;
   hasSearchError: ReplaySubject<boolean>;
-  widget: ReplaySubject<Widget>;
   displayedResource: BehaviorSubject<DisplayedResource>;
   labels: Subject<Labels>;
 };
@@ -45,8 +38,6 @@ let _state: {
   hasSearchError: Observable<boolean>;
   pendingSuggestions: Observable<boolean>;
   pendingResults: Observable<boolean>;
-  widget: Observable<Widget>;
-  customStyle: Observable<string>;
   displayedResource: Observable<DisplayedResource>;
   getMatchingParagraphs: (resId: string) => Observable<Search.Paragraph[]>;
   getMatchingSentences: (resId: string) => Observable<Search.Sentence[]>;
@@ -63,7 +54,6 @@ export const nucliaStore = (): NucliaStore => {
       searchResults: new BehaviorSubject(NO_RESULTS),
       triggerSearch: new Subject(),
       hasSearchError: new ReplaySubject(1),
-      widget: new ReplaySubject(1),
       displayedResource: new BehaviorSubject({ uid: '' }),
       labels: new Subject<Labels>(),
     };
@@ -87,14 +77,6 @@ export const nucliaStore = (): NucliaStore => {
       hasSearchError: _store.hasSearchError.asObservable(),
       pendingSuggestions: _store.suggestions.pipe(map((res) => (res as typeof PENDING_RESULTS).pending)),
       pendingResults: _store.searchResults.pipe(map((res) => (res as typeof PENDING_RESULTS).pending)),
-      widget: _store.widget.asObservable(),
-      customStyle: _store.widget.pipe(
-        map((widget) =>
-          Object.entries(widget?.style || {})
-            .filter(([k, v]) => !!v)
-            .reduce((acc, [k, v]) => `${acc}--custom-${k}: ${v};`, ''),
-        ),
-      ),
       displayedResource: _store.displayedResource.asObservable(),
       getMatchingParagraphs: (resId: string): Observable<Search.Paragraph[]> => {
         return _store!.searchResults.pipe(
