+++ libs/search-widget/src/core/stores/widget.store.ts
@@ -1,0 +1,19 @@
+import { writableSubject } from './svelte-writable-subject';
+import { map, Observable } from 'rxjs';
+import type { WidgetAction } from '../models';
+import type { Widget } from '@nuclia/core';
+
+let widgetActions: WidgetAction[] = [];
+export const setWidgetActions = (actions: WidgetAction[]) => {
+  widgetActions = actions;
+};
+export const getWidgetActions = () => widgetActions;
+
+export const searchWidget = writableSubject<Widget | null>(null);
+export const customStyle: Observable<string> = searchWidget.pipe(
+  map((widget) =>
+    Object.entries(widget?.style || {})
+      .filter(([k, v]) => !!v)
+      .reduce((acc, [k, v]) => `${acc}--custom-${k}: ${v};`, ''),
+  ),
+);
